Add a "prune-local" command to clean up from GH.
This will delete local branches _that have been pushed to a remote_
and that have been merged into master. This will not delete local only
branches that have never been pushed, and does a safe delete "-d", so
if the branch was removed on GH but never merged, it won't delete it.